+++ src/main/java/승재/ch11/P31.java
@@ -1,4 +1,4 @@
-package 승재.ch10;
+package 승재.ch11;
 
 import java.util.HashMap;
 import java.util.Map;
